pyproject.toml
[project]
name = "vata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.11",
    "ddgs>=9.6.1",
    "docling>=2.52.0",
    "dspy-ai>=2.5.1",
    "google-genai==1.19.0",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "manim==0.19.0",
    "manim-voiceover[gtts]==0.3.7",
    "numpy>=1.21.0",
    "pip>=25.2",
    "python-dotenv==1.0.1",
    "setuptools>=80.9.0",
]
[tool.uv]
override-dependencies = [
  "python-dotenv==1.0.1"
]
